(misc) Pass validation for unset optional values
When a parameter is optional and unset, we do not want to validate that
value against the type and validation for the parameter.

A comment indicates that this check is already performed earlier, but it
seems to not be the case in all cases (maybe due to refactoring).

Remove the misleading comment and skip validation if parameter is not
set and is optional, a similar way it is done in
MCollective::Data.ddl_validate()