(misc) Make it easier to work with bundler
While hacking in this gem a user might want to run another project with
their work-in-progress code.  This can be achieved by using something
like the following in that other project's Gemfile:

```ruby
gem 'choria-mcorpc-support', path: '../choria-io/mcorpc-ruby-support'
```

Unfortunately, with the gem specification bundled into the Rakefile,
bundler cannot access it and this result in a failure:

```sh-session
romain@zappy ~/Projects/inventory % bundle
Fetching gem metadata from https://rubygems.org/..
Could not find gem 'choria-mcorpc-support' in source at `../choria-io/mcorpc-ruby-support`.
The source does not contain any versions of 'choria-mcorpc-support'
```

Move the specification in a dedicated Ruby file, and load it from the
Rakefile to prevent this from happening and making it easier to work on
a work-in-porogress gem.