(#149) Improve UX for users of choria.use_srv_records
Choria use `choria.use_srv` to check SRV records while MCollective used
`choria.use_srv_records` for this purpose.  We are updating the mcorpc
support gem to migrate more easily by detecting when a configuration
change is needed, and if so warning the use about it and fixing-it live
when loading the choria configuration.