(#151) Set the environment variable PT__task
When running a task, Bolt sets the name of that task in the `PR__task`
environment variable.  Commands run by Bolt can use this variable to
determine if running in the context of a Bolt task or not.

Also set this variable when running through choria so that code that
wants to behave differently when running from a Bolt task can go through
the same path when executed by the choria tasks support without needing
modification.