(#179) Add a Choria::TaskResult#bolt_task_result API
Bolt tasks can return a Hash or a raw String.  When a raw String is
returned, it is wrapped in a JSON object and the output is placed in an
_output key, so in any case, stdout contains a JSON representation of a
data structure containing the return value of the task.

Add a helper method to easily access to the result of the task
execution.