Fix dubious test
The readlines method of File always return an array (can be empty) if a
file could be read, or raise a Errno::ENOENT if a file does not exist,
Errno::EACCES if it cannot be read, etc.

It seems to never return nil, and the stubbing is not working as
expected with some versions of ruby.  Adjust the test to align with what
would happen in the real world.