(#211) Get dependencies from the gemfile
The mcorpc-ruby-support.gemspec file contains the run dependencies. They
where duplicated in Gemfile, and the Gemfile also included
choria-mcorpc-support, which use the latest published version of the
gem, not the code from the current directory.

Add `gemspec` to the Gemfile and remove unecessary dependencies to avoid
duplication and load the code from the current directory when running
tests.