(misc) Allow latest google-protobuf 3.25.*
Version 3.25.0 has known vulnerability.  While we do not really care
about this development dependency which has no security implication for
us nor our end users, updating it prevent GitHub from warning us that
the project depend on vulnerable code when we push commits.

It is better to have warning when there are actual problems, so that we
do not get used to ignoring a useless warning message and failling to
spot when the number of vulnerabilities increase.